+++ lib/main/stateholder/main_ui_model.dart
@@ -1,5 +1,5 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
-import '../model/label_image.dart';
+import '../../model/label_image.dart';
 part 'main_ui_model.freezed.dart';
 
 @freezed
