+++ lib/main/stateholder/main_ui_model_state_notifier_provider.dart
@@ -1,4 +1,4 @@
-import 'package:ikut_annotation_v2/main/main_ui_model.dart';
+import 'package:ikut_annotation_v2/main/stateholder/main_ui_model.dart';
 import 'package:ikut_annotation_v2/model/label_image.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 part 'main_ui_model_state_notifier_provider.g.dart';
