+++ test/main/stateholder/main_ui_model_provider_test.dart
@@ -1,9 +1,9 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:ikut_annotation_v2/main/main_ui_model.dart';
-import 'package:ikut_annotation_v2/main/main_ui_model_provider.dart';
+import 'package:ikut_annotation_v2/main/stateholder/main_ui_model.dart';
+import 'package:ikut_annotation_v2/main/stateholder/main_ui_model_provider.dart';
 import 'package:ikut_annotation_v2/model/label_image.dart';
 
-import '../provider_container.dart';
+import '../../provider_container.dart';
 
 void main() {
   test('mainUiModelProvider', () {
