+++ test/main/stateholder/main_ui_model_state_notifier_provider_test.dart
@@ -1,10 +1,10 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:ikut_annotation_v2/main/main_ui_model.dart';
-import 'package:ikut_annotation_v2/main/main_ui_model_state_notifier_provider.dart';
+import 'package:ikut_annotation_v2/main/stateholder/main_ui_model.dart';
+import 'package:ikut_annotation_v2/main/stateholder/main_ui_model_state_notifier_provider.dart';
 import 'package:ikut_annotation_v2/model/label_image.dart';
 
-import '../provider_container.dart';
+import '../../provider_container.dart';
 
 void main() {
   group("mainUiModelStateNotifier", () {
